--- src/by/it/kozlov/_tasks_/lesson01/HelloWorld.java
@@ -1,7 +1,0 @@
-package by.it.kozlov._tasks_.lesson01;
-
-public class HelloWorld {
-    public static void main(String[] args) {
-        System.out.println("Hello World!");
-    }
-}
