--- src/by/it/kozlov/_tasks_/lesson02/TaskA1.java
@@ -1,15 +1,0 @@
-package by.it.kozlov._tasks_.lesson02;
-/*
-напишите программу которая выводит на экран
-Hello world!
-
-(строка должа быть одна, ошибки в тексте не допускаются)
-Для ручной проверки запустите программу Ctrl+Shift+F10
-Для автоматической проверки откройте и запустите класс Testing
-*/
-
-class TaskA1 {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
