--- src/by/it/kozlov/_tasks_/lesson02/TaskA2.java
@@ -1,19 +1,0 @@
-package by.it.kozlov._tasks_.lesson02;
-
-/*
-напишите тело main программы которая выводит на экран 5 строк
-"Я начинаю изучать Java!"
-
-Для ручной проверки запустите программу Ctrl+Shift+F10
-Для автоматической проверки откройте и запустите класс Testing
-*/
-
-class TaskA2 {
-    public static void main(String[] args) {
-        System.out.println("Я"+" "+"начинаю"+" "+"изучать"+" "+"Java!");
-        System.out.println("Я"+" "+"начинаю"+" "+"изучать"+" "+"Java!");
-        System.out.println("Я"+" "+"начинаю"+" "+"изучать"+" "+"Java!");
-        System.out.println("Я"+" "+"начинаю"+" "+"изучать"+" "+"Java!");
-        System.out.println("Я"+" "+"начинаю"+" "+"изучать"+" "+"Java!");
-    }
-}
