--- src/by/it/kozlov/_tasks_/lesson02/TaskB3.java
@@ -1,24 +1,0 @@
-package by.it.kozlov._tasks_.lesson02;
-/*
-        Уберите комментарии с части кода, чтобы на экран вывелось сообщение «C Новым Годом».
-
-        Требования:
-        1. Программа должна выводить на экран текст.
-        2. Нужно убрать комментарии в некоторых строках и не менять остальные.
-        3. В программу нельзя добавлять новые строки или удалять имеющиеся.
-        4. Выведенный текст должен соответствовать заданию.
-*/        
-class TaskB3 {
-        public static void main(String[] args) {
-//        String s = "C";
-//        System.out.println("Новым Годом");
-//        System.out.println("C Новым");
-//        System.out.println("Годом");
-        System.out.print("C Новым");
-//        System.out.println("Годом");
-//        System.out.println(s);
-//        System.out.print("C Новым!");
-        System.out.print(" ");
-        System.out.println("Годом");
-    }
-}
