--- src/by/it/kozlov/_tasks_/lesson04/TaskB3.java
@@ -1,31 +1,0 @@
-package by.it.kozlov._tasks_.lesson04;
-/*
-С клавиатуры вводится некоторое число.
-Пусть считается сумма int sum=1+2+3+4+5+... и т.д.
-Каждая цифра суммы выводится в консоль через запятую (без пробела).
-Остановите вывод до того, как сумма превысит введенное число.
-
-Например:
-Ввод:
-122
-Вывод:
-1,3,6,10,15,21,28,36,45,55,66,78,91,105,120,
-
-Ввод:
-15
-Вывод:
-1,3,6,10,15,
-
-Ввод:
-14
-Вывод:
-1,3,6,10,
-
-
- */
-
-
-
-public class TaskB3 {
-
-}
