--- src/by/it/kozlov/_tasks_/lesson04/TaskC2.java
@@ -1,31 +1,0 @@
-package by.it.kozlov._tasks_.lesson04;
-
-/*
-
-На входе вводится число и месяц 2018 года.
-Программа должна выводить номер дня года, считая 1 января первым днем,
-а 31 декабря 365-м.
-
-Пример:
-Ввод:
-1 1
-
-Вывод
-1
-
-Ввод:
-31 12
-
-Вывод
-365
-
-Ввод:
-13 9
-
-Вывод
-256
-
- */
-public class TaskC2 {
-
-}
