+++ src/by/it/kozlov/lesson01/HelloWorld.java
@@ -1,0 +1,7 @@
+package by.it.kozlov.lesson01;
+
+public class HelloWorld {
+    public static void main(String[] args) {
+        System.out.println("Hello World!");
+    }
+}
