+++ src/by/it/kozlov/lesson04/TaskA2.java
@@ -1,0 +1,29 @@
+package by.it.kozlov.lesson04;
+
+/*
+
+День недели
+Ввести с клавиатуры номер дня недели, в зависимости от номера вывести название
+«понедельник», «вторник», «среда», «четверг», «пятница», «суббота», «воскресенье»,
+если введен номер больше 7 или меньше 1 – вывести «такого дня недели не существует».
+
+
+Требования:
+1. Программа должна считывать число c клавиатуры.
+2. Программа должна выводить текст на экран.
+3. Если введено число от 1 до 7, необходимо вывести день недели.
+4. Если введено число не входящее в интервал от 1 до 7,
+    то вывести уведомление что такого дня недели не существует.
+5.  Нельзя использовать массивы или коллекции Java.
+
+Пример для номера 5:
+пятница
+
+Пример для номера 10:
+такого дня недели не существует
+
+ */
+
+public class TaskA2 {
+
+}
